+++ src/page/PokemonRegion.jsx
@@ -1,102 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import '../index.css';
-
-import kantoImage from '../assets/kanto-region-min.png';
-import johtoImage from '../assets/johto-region-min.png';
-import hoenImage from '../assets/hoen-region-min.png';
-import sinnohImage from '../assets/sinnoh-region-min.png';
-import teseliaImage from '../assets/teselia-region-min.png';
-import kalosImage from '../assets/kalos-region-min.png';
-import alolaImage from '../assets/alola-region-min.png';
-import galarImage from '../assets/galar-region-min.png';
-import paldeaImage from '../assets/paldea-region-min.jpg';
 import { X } from 'lucide-react';
 import Boton from '../component/Boton';
-
-const regions = [
-  { id: 'kanto',
-    name: 'Kanto', 
-    image: kantoImage,
-    description: 'Es una región clásica con un diseño basado en Japón central. Aquí comienza la aventura Pokémon original.',
-    professor: 'El Profesor Oak',
-    starterPokemons: ['Bulbasaur', 'Charmander', 'Squirtle'],
-    villain: 'Team Rocket',
-    legendaryPokemons: ['Articuno', 'Zapdos', 'Moltres'],
-  },
-  { id: 'johto',
-    name: 'Johto', 
-    image: johtoImage,
-    description: 'Es una región culturalmente rica inspirada en el Japón occidental.',
-    professor: 'El Profesor Elm',
-    starterPokemons: ['Chikorita', 'Cyndaquil', 'Totodile'],
-    villain: 'Team Rocket',
-    legendaryPokemons: ['Ho-Oh', 'Lugia'],
-  },
-  { id: 'hoenn',
-    name: 'Hoenn', 
-    image: hoenImage,
-    description: 'Una región tropical y diversa, con ecosistemas que van desde selvas hasta desiertos.',
-    professor: 'El Profesor Birch',
-    starterPokemons: ['Treecko', 'Torchic', 'Mudkip'],
-    villain: 'Team Magma y Team Aqua',
-    legendaryPokemons: ['Groudon', 'Kyogre', 'Rayquaza'],
-  },
-  { id: 'sinnoh',
-    name: 'Sinnoh', 
-    image: sinnohImage,
-    description: 'Inspirada en la isla japonesa de Hokkaido, es conocida por su conexión con la mitología y los orígenes del mundo Pokémon.',
-    professor: 'El Profesor Rowan',
-    starterPokemons: ['Turtwig', 'Chimchar', 'Piplup'],
-    villain: 'Team Galactic',
-    legendaryPokemons: ['Dialga', 'Palkia', 'Giratina'],
-  },
-  { id: 'teselia',
-    name: 'Teselia', 
-    image: teseliaImage,
-    description: 'Una región inspirada en Nueva York, con un enfoque urbano y moderno.', 
-    professor: 'La Profesor Juniper',
-    starterPokemons: ['Snivy', 'Tepig', 'Oshawott'],
-    villain: 'Team Plasma',
-    legendaryPokemons: ['Reshiram, Zekrom, Kyurem'],
-  },
-  { id: 'kalos',
-    name: 'Kalos', 
-    image: kalosImage,
-    description: 'Inspirada en Francia, es una región elegante con énfasis en la belleza y la moda.',
-    professor: 'El Profesor Sycamore',
-    starterPokemons: ['Chespin', 'Feenekin', 'Froakie'],
-    villain: 'Team Flare',
-    legendaryPokemons: ['Xerneas', 'Yveltal', 'Zygarde'],
-  },
-  { id: 'alola',
-    name: 'Alola', 
-    image: alolaImage,
-    description: 'Una región inspirada en Hawái, es una región con un sistema único de pruebas en lugar de gimnasios tradicionales.',
-    professor: 'El Profesor Kukui',
-    starterPokemons: ['Rowlet', 'Litten', 'Popplio'],
-    villain: 'Team Skull y Fundación Æther',
-    legendaryPokemons: ['Solgaleo', 'Lunala'],
-  },
-  { id: 'galar',
-    name: 'Galar', 
-    image: galarImage,
-    description: 'Basada en el Reino Unido, mezcla tradición con modernidad.', 
-    professor: 'La Profesor Magnolia',
-    starterPokemons: ['Grookey', 'Scorbunny', 'Sobble'],
-    villain: 'Team Yell y El Presidente Rose',
-    legendaryPokemons: ['Zacian', 'Zamazenta', 'Eternatus'],
-  },
-  { id: 'paldea',
-    name: 'Paldea',
-    image: paldeaImage,
-    description: 'Región inspirada en la Península Ibérica, con paisajes diversos y cultura vibrante.',
-    professor: 'Clavel y Sada/Turo',
-    starterPokemons: ['Sprigatito', 'Fuecoco', 'Quaxly'],
-    villain: 'El Team Star y la Fundación Área Cero',
-    legendaryPokemons: ['Koraidon', 'Miraidon', 'Los cuatro tesoros de la ruina (Wo-Chien, Chien-Pao, Ting-Lu, Chi-Yu)'],
-  }
-]
+import regions from '../js/regions';
 
 function PokemonRegion() {
   const [animatedRegions, setAnimatedRegions] = useState([]);
